Fix: correct the contents for stella PRTS records
correct the contents for the stell PRTS records
