+++ Tools/PC/jira-card-creator/pc_platform_tracker.py
@@ -517,8 +517,78 @@ def generate_platform_tracker(project_name):
     return obj_pj_book
 
 
+def combine_duplicate_tag(data, primary_key):
+    for milestone, platform_data in data.items():
+        needed_data = {}
+        new_data = []
+
+        for platform in platform_data:
+            tag = platform.pop(primary_key)
+            platform_name = platform.pop("platform_name", "")
+            new_name = platform_name.split("(")[0]
+
+            # Put the platform_name into a list even though there's only one
+            # record for this tag
+            # This is for somerville RTS, we will do the enablement tasks for
+            # those systems with the same platform tag
+            if milestone != "rts" and primary_key == "platform_tag":
+                platform.update(
+                    {"platform_name": [new_name], "platform_tag": tag}
+                )
+                new_data.append(platform)
+                continue
+
+            if "product_name" not in platform.keys() or \
+               platform_name == "":
+                continue
+
+            product_name = platform.pop("product_name")
+
+            if tag in needed_data.keys():
+                needed_data[tag]["platform_name"].append(new_name)
+                if product_name is not None:
+                    needed_data[tag]["product_name"].append(product_name)
+            else:
+                platform.update({"platform_name": [new_name]})
+                if product_name is not None:
+                    platform.update({"product_name": [product_name]})
+                needed_data.update({tag: platform})
+
+        # rts only
+        if needed_data:
+            for tag, value in needed_data.items():
+                value.update({primary_key: tag})
+                new_data.append(value)
+
+        data.update({milestone: new_data})
+
+    return data
+
+
+def generate_platform_records(projects):
+    payload = {}
+    for project in projects:
+        if project in ["stella", "sutton"]:
+            primary_key = "lp_tag"
+        else:
+            # by default, we will group record by platform code name
+            primary_key = "platform_tag"
+
+        obj_pj_book = generate_platform_tracker(project)
+        project_payload = obj_pj_book.dump_to_dict("status.eq=in-flight")
+
+        new_payload = combine_duplicate_tag(project_payload, primary_key)
+
+        if new_payload:
+            payload.update({project: new_payload})
+
+    return payload
+
+
 if __name__ == "__main__":
     print(
-        generate_platform_tracker(
-            "sutton").dump_to_json("status.eq=in-flight")
+        json.dumps(
+            generate_platform_records(["somerville", "stella", "sutton"]),
+            indent=4
+        )
     )
